+++ drizzle/0045_fluffy_ender_wiggin.sql
@@ -1,0 +1,2 @@
+ALTER TABLE `notebooks` ADD `shared_with_family` integer DEFAULT false NOT NULL;--> statement-breakpoint
+ALTER TABLE `shopping_categories` ADD `shared_with_family` integer DEFAULT false NOT NULL;